"Added functionality to reset and size button"
"I assigned a function to both the size and reset buttons
that allows the user to click and either be asked how big they
want the grid or to reset the grid. Both kept the value of the
previous grid until i added the bit that set the innerHTML of
gridContainer to nothing. Very simple but works well."