"Adjusted changeColor to accept new variables"
"I had to change the functions attached to blackBtn and rainbowBtn to
arrow functions so that they could accept the parameters I wanted.
I changed the DEFAULT_COLOR to black as well. After adjusting
changeColor I realized that the new conditions made it so that the
user needed to either click rainbowBtn or blackBtn in order for color
to show up when they click on the grid. I added one more 'else if'
in order for black to be the default color, before any button has been
selected."